fix squash check
