+++ core/services/chainlink/application.go
@@ -10,9 +10,6 @@ import (
 	"sync"
 	"syscall"
 
-	"github.com/coreos/go-semver/semver"
-	"github.com/smartcontractkit/chainlink/core/static"
-
 	"github.com/jinzhu/gorm"
 	"github.com/pkg/errors"
 	"github.com/smartcontractkit/chainlink/core/services/fluxmonitorv2"
@@ -124,32 +121,6 @@ type ChainlinkApplication struct {
 	startStopMu sync.Mutex
 }
 
-func CheckSquashUpgrade(db *gorm.DB) error {
-	// Ensure that we don't try to run a node version later than the
-	// squashed database versions without first migrating up to just before the squash.
-	// If we don't see the latest migration and node version >= S, error and recommend
-	// first running version S - 1 (S = version in which migrations are squashed).
-	if static.Version == "unset" {
-		return nil
-	}
-	squashVersionMinus1 := semver.New("0.9.10")
-	currentVersion := semver.New(static.Version)
-	lastV1Migration := "1612225637"
-	if squashVersionMinus1.LessThan(*currentVersion) {
-		// Running code later than S - 1. Ensure that we see
-		// the last v1 migration.
-		q := db.Exec("SELECT * FROM migrations WHERE id = ?", lastV1Migration)
-		if q.Error != nil {
-			return q.Error
-		}
-		if q.RowsAffected == 0 {
-			// Do not have the S-1 migration.
-			return errors.Errorf("Need to upgrade to chainlink version %v first before upgrading to version %v in order to run migrations", squashVersionMinus1, currentVersion)
-		}
-	}
-	return nil
-}
-
 // NewApplication initializes a new store if one is not already
 // present at the configured root directory (default: ~/.chainlink),
 // the logger at the same directory and returns the Application to
@@ -158,9 +129,6 @@ func NewApplication(config *orm.Config, ethClient eth.Client, advisoryLocker pos
 	shutdownSignal := gracefulpanic.NewSignal()
 	store := strpkg.NewStore(config, ethClient, advisoryLocker, shutdownSignal, keyStoreGenerator)
 
-	if err := CheckSquashUpgrade(store.DB); err != nil {
-		panic(err)
-	}
 	setupConfig(config, store)
 
 	explorerClient := synchronization.ExplorerClient(&synchronization.NoopExplorerClient{})
